requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.2
defusedxml==0.7.1
distro==1.9.0
Django==4.2.20
django-allauth==65.8.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
mysqlclient==2.2.7
numpy==2.0.2
oauthlib==3.2.2
openai==1.70.0
openpyxl==3.1.5
pandas==2.2.3
pillow==11.1.0
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.4.0
